move generate method from SchemaGenerator to SchemaGeneratorHelper; bug fix for issue 2055
